+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Fixed `api.ux.datasource.get_to_update()` to actually function with a response o
 Added a step in `driver.ux.driver.sync_picker()` to collapse sequencial whitespace in text input when searching for options on a select type element (drop-downs).  
 This should not be done for text input type pickers (magnifying glass) as they retain the sequencial whitespace in comparisons.
 
+Fixed `api.common.get_response_attribute()` to check if kwarg is single value and compare exact match. List kwargs compare if the attribute is within the set.
+
 ## Added
 
 Added support for ux data source template datetime type recognition.